% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FH_Chi_Quadrat_Eine_Variable.R
\name{FH_Chi_Quadrat_Eine_Variable}
\alias{FH_Chi_Quadrat_Eine_Variable}
\title{Eine Funktion, die bestehende R-Pakete zusammenschnürt um das Ergebnis eines einfachen Chi-Quadrat-Tests mit einer Variablen als Word-Datei auszugeben.}
\usage{
FH_Chi_Quadrat_Eine_Variable(
  data = dataset,
  dv = AV,
  tabnumber = Tabellennummer
)
}
\arguments{
\item{data}{als (data.frame), voreingestellt ist "dataset". Hier muss der Datensatz angegeben werden.}

\item{dv}{als Charakter-Vektor mit der Länge 1. Voreingestellt ist "AV". Hier muss die kategoriale Variable angegeben werden, ueber die der Chi-Quadrat-Test laufen soll.}

\item{tabnumber}{als  Zahl. Voreingestellt ist "Tabellennummer". Das ist die Tabellennummer, die dann in der Tabellenueberschrift der Chi-Quadrat-Tabelle steht.}
}
\value{
Hauptsaechlich wird eine Word-Datei erstellt. Die Word-Datei Word_Tabelle_Haeufigkeitstabelle beinhaltet den Chi-Quadrat-Test.
}
\description{
Eine Funktion, die bestehende R-Pakete zusammenschnürt um das Ergebnis eines einfachen Chi-Quadrat-Tests mit einer Variablen als Word-Datei auszugeben.
}
