% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FH_Chi_Quadrat_Post_hoc_Eine_Variable.R
\name{FH_Chi_Quadrat_Post_Hoc_Eine_Variable}
\alias{FH_Chi_Quadrat_Post_Hoc_Eine_Variable}
\title{Eine Funktion, die bestehende R-Pakete zusammenschn체rt um Post-Hoc-Tests eines einfachen Chi-Quadrat-Tests mit einer Variablen als Word-Datei auszugeben. Diese sollte nur interpretiert werden, wenn der Chi-Quadrat-Test signifikant ist. ACHTUNG: Der Befehl funktioniert nur, wenn die kategoriale Variable mindestens 3 verschiedene Auspraegungen hat.}
\usage{
FH_Chi_Quadrat_Post_Hoc_Eine_Variable(
  data = dataset,
  dv = AV,
  tabnumber = Tabellennummer
)
}
\arguments{
\item{data}{als (data.frame), voreingestellt ist "dataset". Hier muss der Datensatz angegeben werden.}

\item{dv}{als Charakter-Vektor mit der L채nge 1. Voreingestellt ist "AV". Hier muss die kategoriale Variable angegeben werden, ueber die der Chi-Quadrat-Test laufen soll.}

\item{tabnumber}{als Vektor mit der L채nge 1 oder als Zahl. Voreingestellt ist "Tabellennummer". Das ist die Tabellennummer, die dann in der Tabellenueberschrift der Post-Hoc-Tests-Tabelle steht.}
}
\value{
Hauptsaechlich wird eine Word-Datei erstellt. Die Word-Datei Word_Tabelle_Chi_Quadrat_Posthoc beinhaltet die Ergebnisse der Post-Hoc-Tests.
}
\description{
Eine Funktion, die bestehende R-Pakete zusammenschn체rt um Post-Hoc-Tests eines einfachen Chi-Quadrat-Tests mit einer Variablen als Word-Datei auszugeben. Diese sollte nur interpretiert werden, wenn der Chi-Quadrat-Test signifikant ist. ACHTUNG: Der Befehl funktioniert nur, wenn die kategoriale Variable mindestens 3 verschiedene Auspraegungen hat.
}
