% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FH_Chi_Quadrat_Zwei_Variablen.R
\name{FH_Chi_Quadrat_Zwei_Variablen}
\alias{FH_Chi_Quadrat_Zwei_Variablen}
\title{Eine Funktion, die bestehende R-Pakete zusammenschn체rt um das Ergebnis eines einfachen Chi-Quadrat-Tests mit zwei Variablen als Word-Datei auszugeben.}
\usage{
FH_Chi_Quadrat_Zwei_Variablen(
  data = dataset,
  dv = AV,
  iv = UV,
  tabnumber = Tabellennummer
)
}
\arguments{
\item{data}{als (data.frame), voreingestellt ist "dataset". Hier muss der Datensatz angegeben werden.}

\item{dv}{als Charakter-Vektor mit der L채nge 1. Voreingestellt ist "AV". Hier muss die erste kategoriale Variable angegeben werden, ueber die der Chi-Quadrat-Test laufen soll.}

\item{iv}{als Charakter-Vektor mit der L채nge 1. Voreingestellt ist "UV". Hier muss die zweite kategoriale Variable angegeben werden, ueber die der Chi-Quadrat-Test laufen soll.}

\item{tabnumber}{als  Zahl. Voreingestellt ist "Tabellennummer". Das ist die Tabellennummer, die dann in der Tabellenueberschrift der Chi-Quadrat-Tabelle steht.}
}
\value{
Hauptsaechlich wird eine Word-Datei erstellt. Die Word-Datei Word_Tabelle_Haeufigkeitstabelle beinhaltet den Chi-Quadrat-Test.
}
\description{
Eine Funktion, die bestehende R-Pakete zusammenschn체rt um das Ergebnis eines einfachen Chi-Quadrat-Tests mit zwei Variablen als Word-Datei auszugeben.
}
