% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FH_Einfaktorielle_ANOVA.R
\name{FH_Einfaktorielle_ANOVA}
\alias{FH_Einfaktorielle_ANOVA}
\title{Eine Funktion, die bestehende R-Pakete zusammenschnürt um das Ergebnis einer einfaktoriellen Varianzanalyse (ANOVA) als Word-Datei auszugeben.}
\usage{
FH_Einfaktorielle_ANOVA(
  data = dataset,
  dv = AV,
  iv = UV,
  tabnumber = Tabellennummer
)
}
\arguments{
\item{data}{als (data.frame), voreingestellt ist "dataset". Hier muss der Datensatz angegeben werden.}

\item{dv}{als Charakter-Vektor mit der Länge 1. Voreingestellt ist "AV". Hier muss eine numerische Variable (metrisch) angegeben werden.}

\item{iv}{als Charakter-Vektor mit der Länge 1. Voreingestellt ist "UV". Hier muss eine kategoriale Variable (nominal/Faktor, belabeled) angegeben werden.}

\item{tabnumber}{als  Zahl. Voreingestellt ist "Tabellennummer". Das ist die Tabellennummer, die dann in der Tabellenueberschrift der deskriptiven Angaben der ANOVA steht. Der Hypothesentest der ANOVA steht in der ANOVA-Tabelle mit der Tabellennummer tabnumber +1. Die Post-Hoc-Ergebnisse haben dann die Tabellennummer tabnumber+2.}
}
\value{
Hauptsaechlich werden drei Word-Dateien erstellt. Die Word-Datei ANOVA-Tabelle_Tabellennummer_Deskriptive_Ergebnisse_APA. enthält die deskriptiven Angaben der ANOVA, die Word-Datei ANOVA-Tabelle_Tabellennummer_APA"beinhaltet den ANOVA-Hypothesen-Test. Die Word-Datei "Word_Tabelle_ANOVA_Posthoc_Ergebnisse" beinhaltet Homl-Bonferroni-korrigierte Post-Hoc-Tests. Zusaetzlich wird noch eine Abbildung erstellt.
}
\description{
Eine Funktion, die bestehende R-Pakete zusammenschnürt um das Ergebnis einer einfaktoriellen Varianzanalyse (ANOVA) als Word-Datei auszugeben.
}
