% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FH_Itemanalyse.R
\name{FH_Itemanalyse}
\alias{FH_Itemanalyse}
\title{Eine Funktion, die bestehende R-Pakete zusammenschnürt um eine Itemanalyse als Word-Tabelle auszugeben und den Konstruktmittelwert zu errechnen.}
\usage{
FH_Itemanalyse(
  data = dataset,
  variables = Konstrukt_Items,
  min_x,
  max_x,
  tabnumber = Tabellennummer,
  construct_name = "Konstrukt_X"
)
}
\arguments{
\item{data}{als (data.frame), voreingestellt ist "dataset". Hier muss der Datensatz angegeben werden.}

\item{variables}{als Charakter-Vektor mit der Länge >=2. Voreingestellt ist "Konstrukt_Items". Hier muessen alle Variablen angegeben werden, mit denen die Itemanalyse durchgeführt werden soll.}

\item{min_x}{als Zahl. Hier muss die minimal moegliche Itemantwort angeben. Bei einer Skala von 1-5 ist das die 1.}

\item{max_x}{als Zahl. Hier muss die maximal moegliche Itemantwort angeben. Bei einer Skala von 1-5 ist das die 5.}

\item{tabnumber}{als Zahl. Voreingestellt ist "Tabellennummer". Das ist die Tabellennummer, die dann in der Tabellenueberschrift der Itemanalyse steht.}

\item{construct_name}{als Charakter-Vektor mit der Länge 1. Voreingestellt ist "Konstrukt_X". Hier den Namen des Konstruktes angeben, fuer das die Itemanalyse durchgeführt werden soll.}
}
\value{
Hauptsaechlich wird eine Word-Datei erstellt. Die Word-Datei "Word_Tabelle_Itemanalyse.rtf" beinhaltet die Ergebnisse der Itemanalyse. Wenn der Befehl als Objekt gespeichert wird, also bspw. dataset <- FH_Itemanalyse (), dann befindet sich in diesem Datensatzes zusätzlich eine neue Variable, und zwar der Konstruktmittelwert.
}
\description{
Eine Funktion, die bestehende R-Pakete zusammenschnürt um eine Itemanalyse als Word-Tabelle auszugeben und den Konstruktmittelwert zu errechnen.
}
