% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FH_Korrelationsmatrix.R
\name{FH_Korrelationsmatrix}
\alias{FH_Korrelationsmatrix}
\title{Eine Funktion, die bestehende R-Pakete zusammenschnürt um als Ergebnis eine Inter-Korrelations-Matrix auszugeben. ACHTUNG: Alle Korrelationen sind Pearson-Korrelationen.}
\usage{
FH_Korrelationsmatrix(
  data = dataset,
  variables = Variablen_Korrelation,
  tabnumber = Tabellennummer
)
}
\arguments{
\item{data}{als (data.frame), voreingestellt ist "dataset". Hier muss der Datensatz angegeben werden.}

\item{variables}{als Charakter-Vektor mit der Länge X. Voreingestellt ist "Variablen_Korrelation". Hier müssen numerische Variablen (metrisch) angegeben werden.}

\item{tabnumber}{als  Zahl. Voreingestellt ist "Tabellennummer". Das ist die Tabellennummer, die dann in der Tabellenueberschrift der Inter-Korrelations-Matrix steht.}
}
\value{
Es wird ein eine Inter-Korrelationsmatrix als Word-Datei mit dem Namen "Korrelationstabelle_tabnumber_APA ausgeben. Zusätzlich wird noch eine Grafik ausgegeben, diese ist aber nicht APA-konform.
}
\description{
Eine Funktion, die bestehende R-Pakete zusammenschnürt um als Ergebnis eine Inter-Korrelations-Matrix auszugeben. ACHTUNG: Alle Korrelationen sind Pearson-Korrelationen.
}
