% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FH_Lineare_Regression.R
\name{FH_Lineare_Regression}
\alias{FH_Lineare_Regression}
\title{Eine Funktion, die bestehende R-Pakete zusammenschnürt um das Ergebnis einer multiplen Regression in Tabellenform als Word-Datei auszugeben. Zusätzlich wird auch noch eine Interkorrelationstabelle ausgegeben.}
\usage{
FH_Lineare_Regression(
  data = dataset,
  dv = AV,
  iv = UV,
  tabnumber = Tabellennummer
)
}
\arguments{
\item{data}{als (data.frame), voreingestellt ist "dataset". Hier muss der Datensatz angegeben werden.}

\item{dv}{als Charakter-Vektor mit der Länge 1. Voreingestellt ist "AV". Hier muss die AV angegeben werden.}

\item{iv}{als Charakter-Vektor mit der Länge > 1. Voreingestellt ist "UV".Hier müssen die UVs angegeben werden.}

\item{tabnumber}{als Vektor mit der Länge 1 oder als Zahl. Voreingestellt ist "Tabellennummer". Das ist die Tabellennummer, die dann in der Tabellenueberschrift der Inter-Korrelationstabelle steht. Die Tabellennummer der Regressionstabelle ist "tabnumber" + 1-}
}
\value{
Hauptsaechlich werden verschiedene Word-Dateien erstellt. Die Word-Datei "Korrelationstabelle" beinhaltet die Interkorrelationsmatrix, die Word-Datei "Word_Tabelle_multiple_Regression" beinhaltet die Ergebnisse der Regression.
}
\description{
Eine Funktion, die bestehende R-Pakete zusammenschnürt um das Ergebnis einer multiplen Regression in Tabellenform als Word-Datei auszugeben. Zusätzlich wird auch noch eine Interkorrelationstabelle ausgegeben.
}
