% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FH_Logistische_Regression.R
\name{FH_Logistische_Regression}
\alias{FH_Logistische_Regression}
\title{Eine Funktion, die bestehende R-Pakete zusammenschnürt um das Ergebnis einer logistischen Regression (einfach oder multiple) in Tabellenform als Word-Datei auszugeben.}
\usage{
FH_Logistische_Regression(
  data = dataset,
  dv = AV,
  iv = UV,
  tabnumber = Tabellennummer
)
}
\arguments{
\item{data}{als (data.frame), voreingestellt ist "dataset". Hier muss der Datensatz angegeben werden.}

\item{dv}{als Charakter-Vektor mit der Länge 1. Voreingestellt ist "AV". Hier muss die AV angegeben werden.}

\item{iv}{als Charakter-Vektor mit der Länge >= 1. Voreingestellt ist "UV".Hier muss die UV (einfache logistische Regression), oder die UVs (multiple logistische Regression) angegeben werden.}

\item{tabnumber}{als Vektor mit der Länge 1 oder als Zahl. Voreingestellt ist "Tabellennummer". Das ist die Tabellennummer, die dann in der Tabellenueberschrift der Regressionstabelle steht.}
}
\value{
Hauptsaechlich wird eine Word-Datei erstellt. Die Word-Datei "Word_Tabelle_logistische_Regression" beinhaltet die Ergebnisse der Regression.
}
\description{
Eine Funktion, die bestehende R-Pakete zusammenschnürt um das Ergebnis einer logistischen Regression (einfach oder multiple) in Tabellenform als Word-Datei auszugeben.
}
