% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FH_Metrische_Korrelation_Zwei_Variablen.R
\name{FH_Metrische_Korrelation_Zwei_Variablen}
\alias{FH_Metrische_Korrelation_Zwei_Variablen}
\title{Eine Funktion, die bestehende R-Pakete zusammenschn체rt um das Ergebnis einer bivariaten Pearson-Korrelation auszugeben.}
\usage{
FH_Metrische_Korrelation_Zwei_Variablen(data = dataset, dv = AV, iv = UV)
}
\arguments{
\item{data}{als (data.frame), voreingestellt ist "dataset". Hier muss der Datensatz angegeben werden.}

\item{dv}{als Charakter-Vektor mit der L채nge 1. Voreingestellt ist "AV". Hier muss eine numerische Variable (metrisch) angegeben werden.}

\item{iv}{als Charakter-Vektor mit der L채nge 1. Voreingestellt ist "UV". Hier muss eine numerische Variable (metrisch) angegeben werden.}
}
\value{
Es wird ein kurzer Ergebnistext ausgegeben.
}
\description{
Eine Funktion, die bestehende R-Pakete zusammenschn체rt um das Ergebnis einer bivariaten Pearson-Korrelation auszugeben.
}
