% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FH_Ueberblick_Datensatz.R
\name{FH_Ueberblick_Datensatz_sav}
\alias{FH_Ueberblick_Datensatz_sav}
\title{Eine Funktion, die bestehende R-Pakete zusammenschnuert um einen Ueberblick ueber einen geladenen .sav-Datensatz zu geben.}
\usage{
FH_Ueberblick_Datensatz_sav(data = dataset)
}
\arguments{
\item{data}{als (data.frame), voreingestellt ist "dataset". Hier muss der Datensatz angegeben werden.}
}
\value{
in Viewer wird ein Überblick über den Datensatz angezeigt.
}
\description{
Eine Funktion, die bestehende R-Pakete zusammenschnuert um einen Ueberblick ueber einen geladenen .sav-Datensatz zu geben.
}
