% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FH_Z_Transformation.R
\name{FH_Z_Transformation}
\alias{FH_Z_Transformation}
\title{Eine Funktion, die bestehende R-Pakete zusammenschnürt um Variablen zu z-Transformieren}
\usage{
FH_Z_Transformation(data = dataset, variables = Zu_Transformieren)
}
\arguments{
\item{data}{als (data.frame), voreingestellt ist "dataset". Hier muss der Datensatz angegeben werden.}

\item{variables}{als Charakter-Vektor mit der Länge  >= 1. Voreingestellt ist "Zu_Transformieren". Hier muessen die Variablen angegeben werden, die z-transformiert werden sollen.}
}
\value{
z-transformierte Variablen. Der Befehl muss als Objekt gespeichert werden, also bspw. dataset <- FH_Z_Transformation (). Die z-transformierte(n) Variable(n) befinden sich gemeinsam mit allen anderen Variablen Ihres Datensatzes im neu angelegten Objekt (neues oder ueberschriebenes data.frame).
}
\description{
Eine Funktion, die bestehende R-Pakete zusammenschnürt um Variablen zu z-Transformieren
}
